Partial completion of Function revision to rework l2NormOfInteriorJumps().
In this part, we add a new interface to Function to allow arbitrary data, not just solution coefficients, to be associated with owned cells on the mesh.

This passes the test that I wrote for the previous version (i.e., l2NormOfInteriorJumps() still works as before for SimpleSolutionFunction).

Still to do:
1. Add support for the new interface to RepFunction.  (This will also require changes to RieszRep.)
2. Change l2NormOfInteriorJumps() to make it return a map with norms for each MPI-local cell.
3. Change l2NormOfInteriorJumps() to allow the norms to be weighted by the measure of the faces.