Set initial condition in Brio-Wu.
Also added some flux conditions.

Things aren't quite working yet: we do get a solution on first time step, but this has substantial violation of mass and momentum conditions, and we generally don't manage to converge in the nonlinear iteration in the second time step.  Also, for some reason the x-velocity that gets emitted to disk is zero, even at the initial guess, which it should not be...